+++ src/main/java/nextstep/subway/line/ui/LineExceptionHandler.java
@@ -1,12 +1,12 @@
-package nextstep.subway.common.exception;
+package nextstep.subway.line.ui;
 
 import nextstep.subway.line.exception.CreateSectionWithWrongUpStationException;
 import nextstep.subway.line.exception.DeleteSectionWithNotLastException;
 import nextstep.subway.line.exception.DeleteSectionWithOnlyOneException;
 import nextstep.subway.line.exception.DownStationDuplicatedException;
 import nextstep.subway.line.exception.LineNameDuplicatedException;
-import nextstep.subway.section.exception.CannotRemoveRegisteredStation;
-import nextstep.subway.station.exception.StationNameDuplicatedException;
+import nextstep.subway.line.exception.LineNotFoundException;
+import nextstep.subway.section.exception.CannotRemoveRegisteredStationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
@@ -15,15 +15,15 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 @RestControllerAdvice
-public class GlobalExceptionHandler {
+public class LineExceptionHandler {
 
-    @ExceptionHandler(StationNameDuplicatedException.class)
-    public void handle(HttpServletResponse response, StationNameDuplicatedException e) throws IOException {
-        response.sendError(HttpStatus.BAD_REQUEST.value());
+    @ExceptionHandler(LineNotFoundException.class)
+    public void handle(HttpServletResponse response, LineNotFoundException e) throws IOException {
+        response.sendError(HttpStatus.NOT_FOUND.value());
     }
 
-    @ExceptionHandler(CannotRemoveRegisteredStation.class)
-    public void handle(HttpServletResponse response, CannotRemoveRegisteredStation e) throws IOException {
+    @ExceptionHandler(CannotRemoveRegisteredStationException.class)
+    public void handle(HttpServletResponse response, CannotRemoveRegisteredStationException e) throws IOException {
         response.sendError(HttpStatus.BAD_REQUEST.value());
     }
