+++ src/main/java/nextstep/subway/section/exception/CannotRemoveRegisteredStationException.java
@@ -1,8 +1,8 @@
 package nextstep.subway.section.exception;
 
-public class CannotRemoveRegisteredStation extends RuntimeException {
+public class CannotRemoveRegisteredStationException extends RuntimeException {
 
-    public CannotRemoveRegisteredStation(Long id) {
+    public CannotRemoveRegisteredStationException(Long id) {
         super("지하철 구간에 등록되어 있는 지하철역은 제거할 수 없습니다. id: " + id);
     }
 }
